Moving SimConfig to helpers
