+++ frontend/src/Helpers/SimConfig.tsx
